enable dcp saving checkpoint for sharded state dict
